Отказ от связующего модуля. Решение проблемы инициализации
